[ERM-5124] тип свойств, прописанных в DoubleGis.Erm.Platform.Model.Entities.Enums.Infrastructure.PropertiesToEnumMapper, изменен на соответствующие enum'ы
git-tfs-id: [http://uk-tfs02:8080/tfs/DefaultCollection]$/Erm.Core/Dev/Features/EFCF/BLFlex;C25688

+++ Source/ApplicationServices/2Gis.Erm.BLFlex.Operations.Global/MultiCulture/Concrete/Clients/MultiCultureCreateClientByFirmOperationService.cs
@@ -51,7 +51,7 @@ namespace DoubleGis.Erm.BLFlex.Operations.Global.MultiCulture.Concrete.Clients
                     MainFirmId = firm.Id,
                     Name = firm.Name,
                     TerritoryId = firm.TerritoryId,
-                    InformationSource = (int)InformationSource.SalesDepartment,
+                    InformationSource = InformationSource.SalesDepartment,
                     OwnerCode = ownerCode,
                     LastQualifyTime = firm.LastQualifyTime.Value,
                     IsActive = true
